requirements.txt
pyserial==3.5
pygame==2.1.2
natsort==8.1.0